+++ changelog.md
@@ -1,4 +1,4 @@
-### Next version
+### 0.16.11 (2023-12-01)
 * Removed `isFinalError` method from `ErrorInfo` used by `RetryStrategy`, use `isLastAttempt()` instead.
 * Added `RetryInfo` as argument to the `exec` extension function in `RetryStrategy`.
 * Added `retryAttemptException` as an extension property to `org.occurrent.retry.AfterRetryInfo` so that you don't need to use the `getFailedRetryAttemptException` method that returns an `Optional` in the Java interface. Instead, the `retryAttemptException` function returns a `Throwable?`.  Import the extension property from the `org.occurrent.retry.AfterRetryInfoExtensions` file.